requirements.txt
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.12.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.3
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
pyparsing==3.2.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
streamlit==1.39.0
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
